Create Optimized Shortest Path finder
the optimizations we made:

Use heapq instead of PriorityQueue: We replaced the PriorityQueue with a list of tuples that can be used as a heap with heapq. This eliminates the overhead of creating a wrapper object for each item in the queue.

Avoid unnecessary relaxation of vertices: We added a check in the relax method to skip vertices that have already been visited. This can save time if the graph has many edges.

Use a separate boolean list to track visited vertices: We added a new visited list to keep track of visited vertices, instead of using the dist_to list. This makes the code simpler and faster, especially if the graph is dense.

Use a generator to compute the shortest path: We modified the shortest_path method to use a generator expression instead of constructing a list of vertices in reverse order. This saves memory and can be useful if the graph is very large.

Minor code cleanup: We made some minor changes to variable names and formatting to make the code more consistent and easier to read.

Overall, these changes should make the code slightly faster and more memory-efficient for most graphs.